requirements.txt
groq==0.12.0
python-dotenv==1.0.1
streamlit==1.40.1

